Move `Attributes` to core-common
Move `Attributes` from sdk-common to core-common.
