+++ sdk/common/src/main/scala/org/typelevel/otel4s/sdk/common/InstrumentationScope.scala
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk
+package org.typelevel.otel4s
+package sdk
 package common
 
 import cats.Hash
