+++ sdk/trace/src/main/scala/org/typelevel/otel4s/sdk/trace/data/SpanData.scala
@@ -14,7 +14,8 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk
+package org.typelevel.otel4s
+package sdk
 package trace
 package data
