+++ sdk/trace/src/main/scala/org/typelevel/otel4s/sdk/trace/samplers/ParentBasedSampler.scala
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk.trace.samplers
+package org.typelevel.otel4s
+package sdk.trace.samplers
 
-import org.typelevel.otel4s.sdk.Attributes
 import org.typelevel.otel4s.sdk.trace.data.LinkData
 import org.typelevel.otel4s.trace.SpanContext
 import org.typelevel.otel4s.trace.SpanKind
