+++ sdk/trace/src/main/scala/org/typelevel/otel4s/sdk/trace/samplers/SamplingResult.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk.trace.samplers
+package org.typelevel.otel4s
+package sdk.trace.samplers
 
 import cats.Hash
 import cats.Show
 import cats.syntax.show._
-import org.typelevel.otel4s.sdk.Attributes
 import org.typelevel.otel4s.trace.TraceState
 
 /** Sampling result returned by [[Sampler.shouldSample]].
