+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/Cogens.scala
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk
+package org.typelevel.otel4s
+package sdk
 package trace
 
 import org.scalacheck.Cogen
 import org.scalacheck.rng.Seed
-import org.typelevel.otel4s.Attribute
-import org.typelevel.otel4s.AttributeKey
-import org.typelevel.otel4s.AttributeType
 import org.typelevel.otel4s.sdk.common.InstrumentationScope
 import org.typelevel.otel4s.sdk.trace.data.EventData
 import org.typelevel.otel4s.sdk.trace.data.LinkData
