+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/Gens.scala
@@ -14,12 +14,12 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk
+package org.typelevel.otel4s
+package sdk
 package trace
 
 import org.scalacheck.Arbitrary
 import org.scalacheck.Gen
-import org.typelevel.otel4s.Attribute
 import org.typelevel.otel4s.Attribute.KeySelect
 import org.typelevel.otel4s.sdk.common.InstrumentationScope
 import org.typelevel.otel4s.sdk.trace.data.EventData
