+++ sdk/trace/src/test/scala/org/typelevel/otel4s/sdk/trace/samplers/ShouldSampleInput.scala
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package org.typelevel.otel4s.sdk.trace
+package org.typelevel.otel4s
+package sdk.trace
 package samplers
 
 import org.scalacheck.Gen
-import org.typelevel.otel4s.sdk.Attributes
 import org.typelevel.otel4s.sdk.trace.data.LinkData
 import org.typelevel.otel4s.trace.SpanContext
 import org.typelevel.otel4s.trace.SpanKind
